toolbarSettingsService - add setCount method, to replace checkboxClicked
The toolbar needs to know how many items are currently selected to be able to enable/disable items with the `onwhen` attribute.

Previously, this was done by calling checkboxClicked which increments/decrements the counter, but that can sometimes lead to inconsistent state (like toolbarCount going to -2).

Adding a method to simply set the counter, since we already do the counting in miq anyway.