DialogUser - add DialogData#outputConversion
to be called from the ui to get dialogData suitable for API consumption.
Right now, on submit, they just JSON.strigify the actual data and send it to the API.

Which leaves Date instances to auto conversion, and that breaks Date parsing.
Adding an explicit output conversion function and separate ui-classic and ui-service PRs to use it.

https://bugzilla.redhat.com/show_bug.cgi?id=1744413